requirements.txt
FastAPI==0.110.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib==1.7.4
pandas==2.2.1
python-multipart==0.0.9
uvicorn==0.29.0
